# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /library:
    get:
      operationId: listLibraryDocuments
      tags: [library]
      summary: Browse the research library with filters
      parameters:
        - { name: search, in: query, required: false, schema: { type: string } }
        - { name: documentType, in: query, required: false, schema: { type: string } }
        - { name: category, in: query, required: false, schema: { type: string } }
        - { name: university, in: query, required: false, schema: { type: string } }
        - { name: degreeLevel, in: query, required: false, schema: { type: string } }
        - { name: department, in: query, required: false, schema: { type: string } }
        - { name: language, in: query, required: false, schema: { type: string } }
        - { name: tag, in: query, required: false, schema: { type: string } }
        - { name: page, in: query, required: false, schema: { type: integer, minimum: 1 } }
        - { name: pageSize, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 100 } }
      responses:
        "200":
          description: Matching library documents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryListResponse"
...
  /library/{id}:
    get:
      operationId: getLibraryDocument
      tags: [library]
      summary: Get a full library document (with rich content) for preview
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Library document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryDocument"
        "404":
          $ref: "#/components/responses/NotFound"
  /library/{id}/use:
    post:
      operationId: useLibraryDocument
      tags: [library]
      summary: Duplicate a library document into a private editable project copy
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "201":
          description: Created editable copy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          $ref: "#/components/responses/NotFound"
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Validate the admin password and return a session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLogin"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Invalid password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/library:
    get:
      operationId: listAdminLibrary
      tags: [admin]
      summary: List all library documents (including drafts) for administration
      responses:
        "200":
          description: All library documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LibraryDocument"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /admin/library/{id}:
    patch:
      operationId: updateLibraryDocument
      tags: [admin]
      summary: Update library document metadata or content
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LibraryDocumentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryDocument"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteLibraryDocument
      tags: [admin]
      summary: Delete a library document
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "204":
          description: Deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /admin/categories:
    post:
      operationId: createLibraryCategory
      tags: [admin]
      summary: Create a taxonomy entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LibraryCategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryCategory"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /admin/categories/{id}:
    delete:
      operationId: deleteLibraryCategory
      tags: [admin]
      summary: Delete a taxonomy entry
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "204":
          description: Deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /projects/{id}/versions:
    get:
      operationId: listProjectVersions
      tags: [versions]
      summary: List version history for a project
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Versions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DocumentVersion"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      operationId: createProjectVersion
      tags: [versions]
      summary: Snapshot the current project state as a new version
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVersionInput"
      responses:
        "201":
          description: Created version
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentVersion"
        "404":
          $ref: "#/components/responses/NotFound"
  /projects/{id}/versions/{versionId}/restore:
    post:
      operationId: restoreProjectVersion
      tags: [versions]
      summary: Restore a project to a previous version snapshot
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - name: versionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Restored project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  parameters:
    ProjectId:
      name: id
      in: path
      required: true
      schema:
        type: integer
  responses:
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Admin authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Formatting:
      type: object
      properties:
        fontFamily: { type: string }
        fontSize: { type: number }
        headingSize: { type: number }
        subheadingSize: { type: number }
        lineSpacing: { type: number }
        pageSize: { type: string }
        marginTop: { type: number }
        marginBottom: { type: number }
        marginLeft: { type: number }
        marginRight: { type: number }
        paragraphAlign: { type: string }
        firstLineIndent: { type: number }
      required:
        - fontFamily
        - fontSize
        - headingSize
        - subheadingSize
        - lineSpacing
        - pageSize
        - marginTop
        - marginBottom
        - marginLeft
        - marginRight
        - paragraphAlign
        - firstLineIndent
    Section:
      type: object
      properties:
        key: { type: string }
        heading: { type: string }
        content: { type: string }
      required: [key, heading, content]
    Reference:
      type: object
      properties:
        id: { type: string }
        raw: { type: string }
        formatted: { type: string }
        type: { type: string }
        authors: { type: string }
        year: { type: string }
        title: { type: string }
        source: { type: string }
        inTextCitation: { type: string }
      required: [id, formatted]
    VerificationIssue:
      type: object
      properties:
        type: { type: string }
        severity: { type: string }
        message: { type: string }
        location: { type: ["string", "null"] }
        suggestion: { type: ["string", "null"] }
      required: [type, severity, message]
    Verification:
      type: object
      properties:
        readinessScore: { type: integer }
        completeness: { type: integer }
        citations: { type: integer }
        indexing: { type: integer }
        formatting: { type: integer }
        issues:
          type: array
          items:
            $ref: "#/components/schemas/VerificationIssue"
      required: [readinessScore, completeness, citations, indexing, formatting, issues]
    Analysis:
      type: object
      properties:
        researchType: { type: string }
        pageCount: { type: integer }
        language: { type: string }
        specialization: { type: string }
        summary: { type: string }
        detectedSections:
          type: array
          items: { type: string }
      required: [researchType, pageCount, language, specialization, summary, detectedSections]
...
    Project:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        workType: { type: string }
        citationStyle: { type: string }
        language: { type: string }
        templateId: { type: ["integer", "null"] }
        rawContent: { type: string }
        sections:
          type: array
          items:
            $ref: "#/components/schemas/Section"
        references:
          type: array
          items:
            $ref: "#/components/schemas/Reference"
        formatting:
          $ref: "#/components/schemas/Formatting"
        analysis:
          oneOf:
            - $ref: "#/components/schemas/Analysis"
            - type: "null"
        verification:
          oneOf:
            - $ref: "#/components/schemas/Verification"
            - type: "null"
        readinessScore: { type: ["integer", "null"] }
        sourceLibraryDocumentId: { type: ["integer", "null"] }
        richContent: { type: ["string", "null"] }
        layoutMetadata:
          oneOf:
            - $ref: "#/components/schemas/LayoutMetadata"
            - type: "null"
        createdAt: { type: string }
        updatedAt: { type: string }
      required:
        - id
        - title
        - workType
        - citationStyle
        - language
        - rawContent
        - sections
        - references
        - formatting
        - createdAt
        - updatedAt
...
    PageSetup:
      type: object
      properties:
        size: { type: string }
        orientation: { type: string, enum: [portrait, landscape] }
        marginTop: { type: number }
        marginBottom: { type: number }
        marginLeft: { type: number }
        marginRight: { type: number }
    CoverFields:
      type: object
      properties:
        title: { type: string }
        subtitle: { type: string }
        studentName: { type: string }
        supervisor: { type: string }
        university: { type: string }
        faculty: { type: string }
        department: { type: string }
        degree: { type: string }
        year: { type: string }
        logoUrl: { type: string }
    LayoutMetadata:
      type: object
      properties:
        coverPageHtml: { type: string }
        headerHtml: { type: string }
        footerHtml: { type: string }
        showPageNumbers: { type: boolean }
        pageNumberFormat: { type: string }
        pageNumberAlign: { type: string, enum: [left, center, right] }
        pageSetup:
          $ref: "#/components/schemas/PageSetup"
        cover:
          $ref: "#/components/schemas/CoverFields"
    LibraryDocument:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: string }
        documentType: { type: string }
        coverImageUrl: { type: ["string", "null"] }
        university: { type: ["string", "null"] }
        degreeLevel: { type: ["string", "null"] }
        department: { type: ["string", "null"] }
        category: { type: ["string", "null"] }
        language: { type: string }
        tags:
          type: array
          items: { type: string }
        originalFileName: { type: ["string", "null"] }
        originalFileUrl: { type: ["string", "null"] }
        richContent: { type: string }
        layoutMetadata:
          $ref: "#/components/schemas/LayoutMetadata"
        formatting:
          oneOf:
            - $ref: "#/components/schemas/Formatting"
            - type: "null"
        status: { type: string }
        createdAt: { type: string }
        updatedAt: { type: string }
      required:
        - id
        - title
        - description
        - documentType
        - language
        - tags
        - richContent
        - layoutMetadata
        - status
        - createdAt
        - updatedAt
    LibraryDocumentSummary:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: string }
        documentType: { type: string }
        coverImageUrl: { type: ["string", "null"] }
        university: { type: ["string", "null"] }
        degreeLevel: { type: ["string", "null"] }
        department: { type: ["string", "null"] }
        category: { type: ["string", "null"] }
        language: { type: string }
        tags:
          type: array
          items: { type: string }
        status: { type: string }
        createdAt: { type: string }
        updatedAt: { type: string }
      required:
        - id
        - title
        - description
        - documentType
        - language
        - tags
        - status
        - createdAt
        - updatedAt
    LibraryListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/LibraryDocumentSummary"
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
      required: [items, total, page, pageSize]
    LibraryDocumentUpdate:
      type: object
      properties:
        title: { type: string, minLength: 1 }
        description: { type: string }
        documentType: { type: string }
        coverImageUrl: { type: ["string", "null"] }
        university: { type: ["string", "null"] }
        degreeLevel: { type: ["string", "null"] }
        department: { type: ["string", "null"] }
        category: { type: ["string", "null"] }
        language: { type: string }
        tags:
          type: array
          items: { type: string }
        richContent: { type: string }
        layoutMetadata:
          $ref: "#/components/schemas/LayoutMetadata"
        status: { type: string }
...
    LibraryCategory:
      type: object
      properties:
        id: { type: integer }
        kind: { type: string }
        name: { type: string }
        createdAt: { type: string }
      required: [id, kind, name, createdAt]
    LibraryCategoryInput:
      type: object
      properties:
        kind: { type: string }
        name: { type: string, minLength: 1 }
      required: [kind, name]
    AdminLogin:
      type: object
      properties:
        password: { type: string, minLength: 1 }
      required: [password]
    AdminSession:
      type: object
      properties:
        token: { type: string }
      required: [token]
    DocumentSnapshot:
      type: object
      properties:
        title: { type: string }
        richContent: { type: string }
        layoutMetadata:
          $ref: "#/components/schemas/LayoutMetadata"
        formatting:
          oneOf:
            - $ref: "#/components/schemas/Formatting"
            - type: "null"
    DocumentVersion:
      type: object
      properties:
        id: { type: integer }
        projectId: { type: integer }
        label: { type: string }
        snapshot:
          $ref: "#/components/schemas/DocumentSnapshot"
        createdAt: { type: string }
      required: [id, projectId, label, snapshot, createdAt]
    CreateVersionInput:
      type: object
      properties:
        label: { type: string }