# lib/api-zod/src/generated/types/documentVersion.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DocumentSnapshot } from './documentSnapshot';

export interface DocumentVersion {
  id: number;
  projectId: number;
  label: string;
  snapshot: DocumentSnapshot;
  createdAt: string;
}