# lib/api-zod/src/generated/types/layoutMetadata.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CoverFields } from './coverFields';
import type { LayoutMetadataPageNumberAlign } from './layoutMetadataPageNumberAlign';
import type { PageSetup } from './pageSetup';

export interface LayoutMetadata {
  coverPageHtml?: string;
  headerHtml?: string;
  footerHtml?: string;
  showPageNumbers?: boolean;
  pageNumberFormat?: string;
  pageNumberAlign?: LayoutMetadataPageNumberAlign;
  pageSetup?: PageSetup;
  cover?: CoverFields;
}