# lib/api-zod/src/generated/types/libraryDocument.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Formatting } from './formatting';
import type { LayoutMetadata } from './layoutMetadata';

export interface LibraryDocument {
  id: number;
  title: string;
  description: string;
  documentType: string;
  /** @nullable */
  coverImageUrl?: string | null;
  /** @nullable */
  university?: string | null;
  /** @nullable */
  degreeLevel?: string | null;
  /** @nullable */
  department?: string | null;
  /** @nullable */
  category?: string | null;
  language: string;
  tags: string[];
  /** @nullable */
  originalFileName?: string | null;
  /** @nullable */
  originalFileUrl?: string | null;
  richContent: string;
  layoutMetadata: LayoutMetadata;
  formatting?: Formatting | null;
  status: string;
  createdAt: string;
  updatedAt: string;
}