# lib/api-zod/src/generated/types/libraryDocumentSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LibraryDocumentSummary {
  id: number;
  title: string;
  description: string;
  documentType: string;
  /** @nullable */
  coverImageUrl?: string | null;
  /** @nullable */
  university?: string | null;
  /** @nullable */
  degreeLevel?: string | null;
  /** @nullable */
  department?: string | null;
  /** @nullable */
  category?: string | null;
  language: string;
  tags: string[];
  status: string;
  createdAt: string;
  updatedAt: string;
}