# lib/api-zod/src/generated/types/libraryDocumentUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LayoutMetadata } from './layoutMetadata';

export interface LibraryDocumentUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  documentType?: string;
  /** @nullable */
  coverImageUrl?: string | null;
  /** @nullable */
  university?: string | null;
  /** @nullable */
  degreeLevel?: string | null;
  /** @nullable */
  department?: string | null;
  /** @nullable */
  category?: string | null;
  language?: string;
  tags?: string[];
  richContent?: string;
  layoutMetadata?: LayoutMetadata;
  status?: string;
}